HTO-1795: OpenSIPS handles large CSeq values incorrectly

+++ modules/usrloc/ul_mi.c
@@ -140,7 +140,7 @@ static inline int mi_add_aor_node(struct mi_node *parent, urecord_t* r,
 			return -1;
 
 		/* cseq */
-		p = int2str((unsigned long)c->cseq, &len);
+		p = int2str((uint32_t)c->cseq, &len);
 		node = add_mi_node_child( cnode, MI_DUP_VALUE, "Cseq", 4, p, len);
 		if (node==0)
 			return -1;
